Remove BaseActiivty code and keep it empty
