+++ client/src/main/java/com/jraska/github/client/ui/RepoDetailActivity.kt
@@ -26,6 +26,7 @@ class RepoDetailActivity : BaseActivity() {
     super.onCreate(savedInstanceState)
 
     setContentView(R.layout.activity_repo_detail)
+    setSupportActionBar(toolbar)
     repo_detail_recycler.layoutManager = LinearLayoutManager(this)
 
     title = fullRepoName()
