+++ feature-users/src/main/java/com/jraska/github/client/users/UsersModule.kt
@@ -3,15 +3,28 @@ package com.jraska.github.client.users
 import androidx.lifecycle.ViewModel
 import com.jraska.github.client.Config
 import com.jraska.github.client.Navigator
+import com.jraska.github.client.PerApp
 import com.jraska.github.client.analytics.EventAnalytics
 import com.jraska.github.client.rx.AppSchedulers
+
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.ClassKey
 import dagger.multibindings.IntoMap
+import retrofit2.Retrofit
 
 @Module
-object UserViewModelModule {
+object UsersModule {
+  @JvmStatic
+  @Provides
+  @PerApp
+  fun provideUsersRepository(retrofit: Retrofit): UsersRepository {
+    val usersApi = retrofit.create(GitHubUsersApi::class.java)
+    val detailApi = retrofit.create(GitHubUserDetailApi::class.java)
+
+    return GitHubApiUsersRepository(usersApi, detailApi)
+  }
+
   @JvmStatic
   @Provides
   @IntoMap
